refactor(checker): split excess-property helpers out of state_checking/property.rs
property.rs (state_checking) sits at its 1480 size ratchet on main, so any
touch trips it (it gates #15449 and future property-checking work). Extract the
ten excess-property diagnostic helpers that precede the
check_object_literal_excess_properties monolith into a new sibling
property/excess_property_helpers.rs (pub(crate), matching the existing
property/ submodules). Pure code motion; no behavior change. The now-unused
syntax_kind_ext import (its uses moved with the cluster) and a stale HashSet
import are dropped from property.rs.

property.rs 1480->1246; new sibling ~248. No other file affected.
